add test cases to MeasureUnitTest
